### 2023-10-25 af78dbc
README created for the jointVIP simulation folder
