+++ paper/simulation/README.txt
@@ -1,0 +1,10 @@
+The simulation in this particular folder refers to the paper 
+"Using Joint Variable Importance Plots to Prioritize Variables 
+in Assessing the Impact of Glyburide on Adverse Birth Outcomes"
+that can be found on arXiv: https://arxiv.org/abs/2301.09754
+
+This is a separate section from the jointVIP code introduction
+paper in the simulation folder.
+This code is a gentle walk through of how one might use the
+jointVIP and how it reduces the point estimate and its variance.
+In the setting of small treated group with a constant effect.
