### 2023-11-27 #6093 @editorialbot
[REVIEW]: jointVIP: Prioritizing variables in observational study design with joint variable importance plot in R
**Submitting author:** @ldliao (<a href="http://orcid.org/0000-0003-4697-6909">Lauren Liao</a>)
**Repository:** https://github.com/ldliao/jointVIP
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.2
