### 2023-12-05 ldliao/jointVIP#2 @jackmwolf
JOSS Review: @jackmwolf
Hi @ldliao! I'm done with my initial review of jointVIP for [openjournals/joss-reviews/issues/6093](https://github.com/openjournals/joss-reviews/issues/6093). The functionality and purpose are great overall. I've listed several minor suggestions for improvement below:

1.   Based on the provided examples and my own testing, it appears that `create_jointVIP()` does not support categorical variables with \>2 levels and such variables must be converted into a set of indicator variables first. I suggest being more explicit about the input requirements in the documentation for `create_jointVIP()`. Currently, the only listed requirements are that `pilot_df` and `analysis_df` are `data.frames`.

1.  Similarly, the `create_jointVIP()` returns an error if `treatment` is not binary (e.g., if it is coded as the strings `'treatment'` and `'control'`). This expectation could be made explicit in the documentation.

1.   Consider adding documentation for how others can contribute to your software somewhere in your root directory (e.g., `CONTRIBUTING.md`). You can find an [example here](https://contributing.md/example/).

1. The GitHub repository contains both `LICENSE` and `LICENSE.md`. The former only lists the year and copyright holder and is not a software license.

1. Can you expand on the statement that "[b]ias curves enable comparisons to support prioritization" in `README.md`? What is prioritized?

1.   (Very minor) Line 62 of `paper.md` says "functionto" instead of "function to."
