### 2024-01-09 58f4c7b
added one-hot encoding and other specifications as requested by reviewer
