+++ paper/paper.md
@@ -59,7 +59,7 @@ library(jointVIP)
 
 To create an object of the jointVIP class, the user needs to supply two datasets and specify the treatment, outcome, and background variable names. Two processed datasets,“pilot” and “analysis” samples, are in the form of data.frames. The analysis sample contains both treated and control groups. The pilot sample contains only control individuals, and they are excluded from the subsequent analysis stage. The treatment variable must be binary: 0 specified for the control group and 1 specified for the treated group. Background variables are measured before both treatment and outcome. The outcome of interest can be either binary or continuous.  
 
-We demonstrate the utility of this package to investigate the effect of a job training program on earnings [@causaldata; @dehejia1999causal; @lalonde1986evaluating].  The treatment is whether the individual is selected for the job training program. The outcomeis earnings in 1978. Covariates are age, education, race/ethnicity, and previous earnings in 1974 and 1975.  After [preprocessing both dataset and log-transforming the earnings,](https://cran.rstudio.com/web/packages/jointVIP/vignettes/jointVIP.html), we use the `create_jointVIP()` functionto create a jointVIP object stored as *new_jointVIP*. 
+We demonstrate the utility of this package to investigate the effect of a job training program on earnings [@causaldata; @dehejia1999causal; @lalonde1986evaluating].  The treatment is whether the individual is selected for the job training program. The outcomeis earnings in 1978. Covariates are age, education, race/ethnicity, and previous earnings in 1974 and 1975.  After [preprocessing both dataset and log-transforming the earnings,](https://cran.rstudio.com/web/packages/jointVIP/vignettes/jointVIP.html), we use the `create_jointVIP()` function to create a jointVIP object stored as *new_jointVIP*. 
 
 
 
