### 2024-01-12 6eff410
updating the licensing information to be more consistent with how r package is set up
