### 2024-01-20 ldliao/jointVIP#2 @jackmwolf
Hi again, @ldliao! All of these changes look great, I'm happy with the state of the software and paper and will update my review in  https://github.com/openjournals/joss-reviews/issues/6093. Please merge this branch into main before issuing the package release  that will be included with the JOSS paper. Thanks!
