### 2024-01-21 ldliao/jointVIP#3 @nhejazi
JOSS review
## Code/Repo

- Note that the example in the README doesn't work immediately when copy-pasted into an R session (I often do this, presumably others too) because the object `df` used in the example is created in a hidden `Rmd` code block, which downloads a CSV of a dataset that is then cleaned to produce `df`. A cleaner solution would be to move this hidden block to a standalone `R` script (stored in the package) and distribute the cleaned dataset as part of your package, as described at https://r-pkgs.org/data.html; this has the advantage of making the package self-contained wrt data, as changes in the availability of that CSV won't affect this package's examples/demos.
- Code organization: Most of the package's core functions seem to be organized in a file `general.R`, which is fine but may prove challenging to work with for contributors. It's recommended that functions be stored in individual files or at least thematically organized (e.g., `classes.R` to store your S3 class definitions); see https://r-pkgs.org/code.html#sec-code-organising.
- The suite of unit tests seems to be reasonably diverse but it's hard to get a sense without a code coverage check in place. On a related note, GitHub Actions seems to be activated but is only being used to produce the JOSS PDF. A better use of the Actions workflow would be to automate the running of unit tests (activated on push and PRs) and code coverage checks; some examples of actions for `R` are at https://github.com/r-lib/actions
- Per JOSS review guidelines, some statement about community guidelines is necessary. This could be either as a separate markdown file in the repo or a new section in the existing README file.

## Paper

- There's a minor typo in the abstract (line 14 of JOSS PDF): "The joint variable importance plot (jointVIP) *package to guides decisions* about which variables to prioritize for adjustment by quantifying and visualizing each variable's relationship to both treatment and outcome" (emphasis mine).
- On "Development" (lines 49-50 of JOSS PDF): It would be clearer to note that the package exposes a new S3 class, called `jointVIP`, and exposes methods of the generic functions `print()`, `summary()`, and `plot()` for this class (as opposed to saying that it "leverages system generic functions").
- On "Usage", a few typos: "outcomeis" instead of "outcome is" and "functionto" instead of "function to"
