### 2024-01-29 d143b69
updated news after merging one_hot branch
