### 2024-01-29 ldliao/jointVIP#2 @ldliao
Hi @jackmwolf thanks for taking the time for the review! I just merged the one_hot branch into main (and updated the news to reflect the latest updates!
