### 2024-02-16 #6093 @nhejazi
@ajstewartlang I should have commented here but I filed a couple weeks ago an issue with the package repo (https://github.com/ldliao/jointVIP/issues/3) with comments from my review. Once addressed, I think these would complete the items currently left unchecked in my reviewer checklist above.
