### 2024-03-04 ldliao/jointVIP#3 @ldliao
@nhejazi apologies for the delay in response!

Thank you so much for taking the time to review.
The reorganization of the code will be taking some more time; but the edits in [paper](https://github.com/ldliao/jointVIP/blob/paper_edits/paper/paper.pdf) is completed (as seen below). Once approved, I'll go ahead and merge the paper_edits branch with the main branch.

**Paper**
- There's a minor typo in the abstract (line 14 of JOSS PDF): "The joint variable importance plot (jointVIP) package to guides decisions about which variables to prioritize for adjustment by quantifying and visualizing each variable's relationship to both treatment and outcome" (emphasis mine).
   - Thank you for highlighting this typo. The updated text is below.
   - The joint variable importance plot (jointVIP) package guides decisions about which variables to prioritize for adjustment by quantifying and visualizing each variable's relationship to both treatment and outcome.

- On "Development" (lines 49-50 of JOSS PDF): It would be clearer to note that the package exposes a new S3 class, called jointVIP, and exposes methods of the generic functions print(), summary(), and plot() for this class (as opposed to saying that it "leverages system generic functions").
   - Thank you for pointing out this for clarification. The updated text is below.
   - The `jointVIP` package was created in the R programming language. The package creates a new S3 class called "jointvip" and uses S3 generic to dispatch `print()`, `summary()`, and `plot()`.

- On "Usage", a few typos: "outcomeis" instead of "outcome is" and "functionto" instead of "function to"
   - Thank you for pointing out these typos. They have been updated accordingly.
