### 2024-04-15 ldliao/jointVIP#3 @nhejazi
hi @ldliao, please see a recent comment in https://github.com/openjournals/joss-reviews/issues/6093 about the status of this review. my understanding is that some of the changes to the code are still outstanding, so my JOSS review needs to stay open, but please let me know if the changes have been made and i'll then check off the relevant items
