### 2024-04-15 #6093 @nhejazi
@ajstewartlang my understanding from @ldliao's comment in https://github.com/ldliao/jointVIP/issues/3 is that some of the review items (changes to the code) remain outstanding, so that i'll need to evaluate the completion of said items in order for my review to be considered complete by JOSS
