### 2024-09-11 #6093 @nhejazi
i've now had a chance to review the changes made in response to my prior review comments (as noted at https://github.com/ldliao/jointVIP/issues/3#issuecomment-2344101918) and do not have any further comments to add. the package looks good! i've checked off the remaining items in my reviewer checklist above
