### 2024-10-23 #6093 @ajstewartlang
@ldliao This looks like the correct DOI for the causaldata package:

- 10.32614/cran.package.causaldata may be a valid DOI for title: causaldata: Example Data Sets for Causal Inference...

Can you check it's indexing the correct version of the package and update the reference list if so?

For MatchIt, I believe the correct DOI is 10.18637/jss.v042.i08

Can you add that please?

Also, there is at least one paper in the `paper.bib` file that doesn't appear in `paper.md`. Could you check this please and ensure that all entries in `paper.bib` are cited in the paper:

@article{benmichael2021,
  title={The Balancing Act in Causal Inference},
  author={Ben-Michael, Eli and Feller, Avi and Hirshberg, David A and Zubizarreta, Jos{\'e} R},
  journal={arXiv preprint arXiv:2110.14831},
  year={2021}
}

Thanks!
