### 2024-10-25 #6093 @ldliao
Hi @ajstewartlang I have updated the paper bib as you requested. Many thanks!
