### 2024-11-01 #6093 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
