### 2024-11-12 #6093 @samhforbes
Hi @ldliao can you set the licence of the Zenodo archive to match the repository please (I think you currently have cc-by rather than MIT).
