### 2024-11-12 #6093 @samhforbes
Oops no I see the old one is the permanent DOI
